Copy file จาก develope

+++ src/app/pages/e-vr/custom-action/custom-action.component.ts
@@ -1,36 +1,36 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { VrService } from '../../../@core/shared/services/vr.service';
-import { NbDialogRef, NbDialogService } from '@nebular/theme';
-import { DeleteConfirmationComponent } from '../delete-confirmation/delete-confirmation.component';
+import { Component, Input, OnInit } from "@angular/core";
+import { VrService } from "../../../@core/shared/services/vr.service";
+import { NbDialogRef, NbDialogService } from "@nebular/theme";
+import { DeleteConfirmationComponent } from "../delete-confirmation/delete-confirmation.component";
 
 @Component({
-  selector: 'ngx-custom-action',
-  templateUrl: './custom-action.component.html',
-  styleUrls: ['./custom-action.component.scss']
+  selector: "ngx-custom-action",
+  templateUrl: "./custom-action.component.html",
+  styleUrls: ["./custom-action.component.scss"],
 })
 export class CustomActionComponent implements OnInit {
   dialogRef: NbDialogRef<DeleteConfirmationComponent>;
   @Input() rowData: any;
 
-  constructor(private vrService: VrService, private dialogService: NbDialogService) { 
-  }
+  constructor(
+    private vrService: VrService,
+    private dialogService: NbDialogService,
+  ) {}
 
-  ngOnInit(): void { }
+  ngOnInit(): void {}
 
   deleteVR(): void {
-    this.dialogRef = this.dialogService.open(DeleteConfirmationComponent,{
+    this.dialogRef = this.dialogService.open(DeleteConfirmationComponent, {
       context: {
         data: {
-          title: 'Do you want to delete this VR?',
-          information: 'You try to delete this VR.'
-        }
-      }
+          title: "Do you want to delete this VR?",
+          information: "You try to delete this VR.",
+        },
+      },
     });
-    
-    this.dialogRef.onClose.subscribe(confirm => {
-      if(confirm == 'affirm')
-        this.vrService.deleteVR(this.rowData.uuid);
+
+    this.dialogRef.onClose.subscribe((confirm) => {
+      if (confirm == "affirm") this.vrService.deleteVR(this.rowData.uuid);
     });
   }
-
 }
